Handle mixers that don't have a playback switch
Some USB soundcards don't provide a master volume. In this case, it is
possible to use an ALSA plugin, 'softvol', that emulates volume in
software.
However, the softvol device lacks a playback switch, and appears as muted
all the time in pnmixer.

This commits fixes that, so that the device appears unmuted, and
clicking on the mute button has no effect.