Fix vol_scroll_event callback
This function used to get the new value of the GtkRange by asking it to
the GtkAdjustment associated. But the fact is that, when we're executing
this callback, the GtkAdjustment value has not been updated yet, so this
call always returned the previous value.

The bug is easy to spot: just open the volume popup, and click anywhere
on the bar. The slider moves there, but you can here that the volume
doesn't change.

The solution is to get the new value that is directly given in the
arguments of the callback. The only minor trouble with that: we must
manually ensure that the value is inside the boundaries defined by the
GtkAdjustment.