Don't hide volume popup on double-click #73 #74
Somehow, a double or triple click causes the hide_me() callback to be
invoked. I guess it's because the button-press-event is connected to it.

Anyway, we don't want to hide the popup on double-click (#73).
Furthermore, with gtk2, the vol_scroll_event() callback is called forever
in case the popup is closed. It's a bug (#74).