Fix setting volume by small step
If volume change step is small (e.g. set to 1) it might not be possible to
set all values - driver restriction. This might lead to a behavior where
the volume can not be changed any more.

Changes made in this commit will prevent such a situation by allowing driver
to set first bellow or first above value according to the change direction.