Change open/close_mixer API
The interger returned was not used anyway. Things look easier like this.

Furthermore, if you look at 21addad, the 'mixer' argument in open_mixer()
is set to NULL on failure, and other part of the code check for that.
You may also notice that 'mixer' is not set to NULL in close_mixer(),
which is a brand hew bug I just introduced in 21addad.

What to think about that ? Appart from the fact that I'm a poor coder,
it also shows that the previous API of these two functions was slightly
more complicated than it was needed, and complexity is always
error-prone.

Let's simplify !