Polishing text notification
Set notification translatable.

Remove app name from notification body.

I think we shouldn't explicitely set the app name. App name is given in
notify_init() anyway, so it' s up to the notification daemon to decide
whether to display application name or not.

Notice that we could provide PNMixer icon with the text notification,
except that PNMixer icons are not installed in the standard icon
directory, therefore only PNMixer can use it. Maybe we should change
that.