Handle missing 'audio-volume-off' icon
If not available, we replace it by 'audio-volume-low', it's the icon
that was used before introducing 'audio-volume-off'.

It's better to remove the 'report_error()' inside get_stock_pixbuf(),
since the error is likely to happen, and it's an error that we handle
nicely. No need to freak out the user.