Replace selected_card_dev() by selected_card_acard()
The reason comes from the code in alsaset(). At some point we open the
selected card, simply by getting the selected card name and card dev,
without any other checks.

It may happen that the selected card has no channels available (due to
changes in alsa conf, in my case). This case goes undetected at this
moment, but leads to an assert() failure later on.

With this change, we get the acard struct from the selected card name,
which permits to check that the sound card is OK (that is, it's part of
the available cards, and it has active channels).