Do not let io_watches multiply
Before this commit, io watches were never removed. As indicated in the
GLib IO Channels documentation, GIOChannels are created with a reference
count of 1, and calling g_io_add_watch() increment by one. As a
consequence, simply calling g_io_unref() is not enough to remove the io
watch. This can easily be seen by putting a debug message in poll_cb(),
then reloading alsa.

Alright, to fix that we must remember the io watch id, and explicitely
remove it in alsaunset(). Also, because the number of watch id may be
more than one, we need an array. I have no idea of the maximum number of
watch ids, I set it to 8. Time will tell.