Fix on_card_changed()
In the previous implementation, we assumed that the card index in the
name was the sae as the card index in the combobox. This was wrong,
since only cards with playable channels are added to the combobox.

Maybe we should just discard cards without channels from the list from
the beginning, because it's the second bug I find related to that.