Fix fill_card_combo()
Before this commit, this function used to match the selected card
against the list of cards. In case selected card is not available
(because unplugged), nothing matched, In this case, the code assumed
that the active card was the first card of the list. And this was the
problem, since the first card of the list (which is always default) is
not always available. This ended up in mixed up things, the card name
in the combo box didn't match the channels combo box.

TO fix that, we need to know which card is ACTIVE, not which one is
SELECTED. Then the problem solves iself.