Split popup ui file in two files
Makes sense since the two popup windows are completely unrelated.

So we know have one file for the menu popup, and one file for the volume
popup.

The file for volume popup is still named popup-window, because renaming
it involves a lot of renaming in source files, and this is not what this
commit is about. But I think it should be renamed in popup-volume in a
near future, it would make more sense.