Rework Gtk2 volume popup to match the gtk3 version
So, mainly, reomve the pictures on left and right of the slider.

It's worth to notice that the pictures were supposed to represent '-' and
'+'. The stock icons used for that were 'gtk-remove' and 'gtk-add'. On
my icon theme, 'gtk-remove' is represented with a cross, not a dash, so
the picture was quite out of place.

So removing the icons is not a bad idea.