Fix vol meter width on startup
Currently it's bogus at startup, because update_status_icons() is called
once before the tray icon exist, leading to a wrong calculation of the
volume meter width. Then it's called a second time when the tray icon is
created, and this time the width is calculated right, but the icon is not
updated since it already exists, and the code thinks nothing changed.

To solve that, just recreate the volume meter array each time, don't try
to be clever and check if it's really needed.