Save colors as double, no matter which Gtk version
At the moment color meter is bogus in Gtk2, and the code is overly
complicated. In the config file, colors are saved as integer for Gtk2,
and as double for Gtk3. There's no need to do that, the config file
shouldn't depend on the Gtk version.

To simplify it, I choose to save colors as double (between 0 and 1).
This is what is done in Gtk3 version.

For Gtk2, we just need to change the values (from double to integer)
when we deal with the UI and GckColor.