build: install a pre-commit hook in autogen
This is basically the default git pre-commit hook, with our additional
checks at the end. At the moment, our additional checks are limited to
checking the date in the man page in case it's changed.

It would have been nicer if the file would contain only our addition,
and we would append it at the end of the default git hook. However it's
not possible because the default git hook ends with an `exec` statement,
so anything appended afterward won't be executed.