Fix unsafe call to 'hotkey_matches' (thx to 00cpxxx) #159
We should always check that an object exist before invoking a method on
it. That's what we do now in `hotkeys.c`.

We can use `g_return_if_fail()` to display an error message in case of
improper invokation of a method with a NULL pointer. It's better than a
segfault I guess.