hacking: re-write the translation part
Now that we use gettext instead of intltool, we version the POT file,
and therefore there's no need to create an archive manually anymore.

Furthermore, I now have a better understanding of the right workflow to
have with the TP. So I took time to describe it with details. In the end
it's a very simple and straightforward workflow.